requirements.txt
fastapi
uvicorn
requests
PyGithub
openai
python-dotenv
